extract fixtures/bootstrap.py
