+++ integration-testing/src/fixtures/common.py
@@ -4,8 +4,6 @@ import pytest
 import random
 import tools.random
 import tempfile
-from tools.rnode import create_bootstrap_node, create_peer_nodes
-from tools.wait import wait_for, string_contains, node_logs, network_converged
 from tools.util import log_box
 from tools.profiling import log_prof_data
 import tools.resources as resources
@@ -100,22 +98,3 @@ def validators_data(config):
 
     with create_bonds_file(validator_keys) as bonds_file:
         yield (bonds_file, validator_keys[0], validator_keys[1:])
-
-
-
-@pytest.fixture(scope="package")
-def bootstrap(docker, docker_network, config, validators_data):
-    bonds_file, bootstrap_keys, _ = validators_data
-    node = create_bootstrap_node(docker, docker_network, bonds_file, bootstrap_keys, config.rnode_timeout)
-
-    yield node
-
-    node.cleanup()
-
-@pytest.fixture(scope="package")
-def started_bootstrap(config, bootstrap):
-    wait_for( string_contains( node_logs(bootstrap),
-                               "coop.rchain.node.NodeRuntime - Listening for traffic on rnode"),
-              config.node_startup_timeout,
-              "Bootstrap node didn't start correctly")
-    yield bootstrap
