+++ integration-testing/src/fixtures/complete_connected.py
@@ -5,6 +5,7 @@ from tools.rnode import create_bootstrap_node, create_peer_nodes
 from tools.wait import wait_for, string_contains, node_logs, network_converged
 from fixtures.datastructures import RChain
 from fixtures.common import *
+from fixtures.bootstrap import *
 
 @pytest.fixture(scope="package")
 def complete_network(config, docker, started_bootstrap, docker_network, validators_data):
